+++ packages/ilmomasiina-backend/src/i18next.d.ts
@@ -3,5 +3,6 @@ import { i18nResources } from "./i18n";
 declare module "i18next" {
   interface CustomTypeOptions {
     resources: (typeof i18nResources)["fi"];
+    allowObjectInHTMLChildren: true;
   }
 }