+++ packages/ilmomasiina-backend/src/mail/config.ts
@@ -1,19 +1,33 @@
-import nodemailer, { Transporter } from "nodemailer";
+import Mailgun from "mailgun.js";
+import nodemailer, { Transport } from "nodemailer";
 import SMTPTransport from "nodemailer/lib/smtp-transport";
-import mailgun from "nodemailer-mailgun-transport";
 
 import config from "../config";
 
-const mailTransporter: Transporter = (() => {
-  if (process.env.NODE_ENV === "test") {
+/** Common input values valid for both nodemailer and mailgun.js. */
+type AbstractMailOptions = {
+  to: string;
+  from: string;
+  subject: string;
+  html: string;
+};
+
+/** Base type for the minimal interface we use in nodemailer. */
+type AbstractTransporter = {
+  sendMail: (options: AbstractMailOptions) => Promise<unknown>;
+  transporter: Pick<Transport, "name">;
+};
+
+const mailTransporter: AbstractTransporter = (() => {
+  if (config.nodeEnv === "test") {
     return nodemailer.createTransport({
       name: "console fallback",
       version: "0",
       send(mail, callback) {
-        // Ignore emails in test environment
         const { message } = mail;
         const envelope = message.getEnvelope();
         const messageId = message.messageId();
+        // Completely ignore emails in test environment - mocking is done at EmailService.send before calling this
         setImmediate(() => callback(null, { envelope, messageId } as any));
       },
     });
@@ -23,15 +37,17 @@ const mailTransporter: Transporter = (() => {
     if (!config.mailgunDomain) {
       throw new Error("Invalid email config: MAILGUN_DOMAIN must be set with MAILGUN_API_KEY.");
     }
-    return nodemailer.createTransport(
-      mailgun({
-        auth: {
-          api_key: config.mailgunApiKey,
-          domain: config.mailgunDomain,
-        },
-        host: config.mailgunHost,
-      }),
-    );
+    const mailgun = new Mailgun(FormData);
+    const client = mailgun.client({
+      username: "api",
+      key: config.mailgunApiKey,
+      url: config.mailgunHost ? `https://${config.mailgunHost}` : undefined,
+    });
+    // Wrap mailgun.js client to somewhat match nodemailer interface
+    return {
+      sendMail: (msg) => client.messages.create(config.mailgunDomain!, msg),
+      transporter: { name: "Mailgun" },
+    };
   }
 
   if (config.smtpHost) {
@@ -46,7 +62,7 @@ const mailTransporter: Transporter = (() => {
         user: config.smtpUser,
         pass: config.smtpPassword,
       },
-    } as SMTPTransport.Options);
+    } satisfies SMTPTransport.Options);
   }
 
   console.warn("Neither Mailgun nor SMTP is configured. Falling back to debug mail service.");